Create storm_May2022.R
Analysis of hourly wind speed, wind direction, and barometric pressure to characterize a storm event in May 2022. Furthermore, the Beaufort wind force scale was applied to the wind speed to further characterize the wind conditions of the VCR. The meteorological data was provided by NOAA tide monitoring station Wachapreague (site no: 8631044). This analysis produced Figure 5 and SI Figure 1 in the manuscript.